+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 11. Terminal inside editor
 12. Theme changer
 13. Added tags in suggestions
+14. Snippet code interface
 
 ### Fixed/Changes: